requirements.txt
awscrt==0.11.4
awsiotsdk==1.5.10
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
